+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.11.0] 2025-07-14
+
+### Added
+
+- GUI_tools: `FeatureMaster`, `GUIHandler` to fully handle GUI-model communication.
+
+### Fixed
+
+- ensemble_learning: `InferenceHandler` uses constant keys for classification outputs.
+
+### Deleted
+
+- GUI_tools: `update_target_fields()` and `BaseFeatureHandler` in favor of the new fully automated OOP approach
+
 ## [3.10.2] 2025-07-12
 
 ### Fixed